Move Email class into Services directory
